% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{as_variable_slots}
\alias{as_variable_slots}
\title{Turn one role argument into a named list of slots}
\usage{
as_variable_slots(value, role, call = rlang::caller_env())
}
\description{
Turn one role argument into a named list of slots
}
\keyword{internal}
