% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{coerce_to_pair_df}
\alias{coerce_to_pair_df}
\title{Coerce links to a \code{from}/\code{to} table}
\usage{
coerce_to_pair_df(x)
}
\arguments{
\item{x}{A data frame with \code{from} and \code{to}, or a list of \code{c(from, to)}
pairs.}
}
\value{
A tibble with character \code{from} and \code{to}.
}
\description{
Coerce links to a \code{from}/\code{to} table
}
\keyword{internal}
