% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_default.R
\name{default_metadata_leaf}
\alias{default_metadata_leaf}
\title{The default value of one flat-addressable metadata field}
\usage{
default_metadata_leaf(field)
}
\arguments{
\item{field}{Length-one character.}
}
\value{
The field's default value, or \code{NULL} for an unknown field.
}
\description{
The default value of one flat-addressable metadata field
}
\keyword{internal}
