% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angles.R
\name{deg_to_rad}
\alias{deg_to_rad}
\title{Convert degrees to radians}
\usage{
deg_to_rad(x)
}
\arguments{
\item{x}{Numeric vector of angles (degrees).}
}
\value{
Numeric vector of angles expressed in radians.
}
\description{
Convert degrees to radians
}
\examples{
deg_to_rad(180)
deg_to_rad(c(0, 90, 180))

}
\seealso{
Other angle utilities:
\code{\link[=angle_to_rad]{angle_to_rad()}},
\code{\link[=rad_to_deg]{rad_to_deg()}},
\code{\link[=unwrap_angle]{unwrap_angle()}},
\code{\link[=wrap_angle]{wrap_angle()}}
}
\concept{angle utilities}
