% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anipoint_validate.R
\name{ensure_has_declared_variables}
\alias{ensure_has_declared_variables}
\title{Ensure every declared variable names a column that exists}
\usage{
ensure_has_declared_variables(data)
}
\arguments{
\item{data}{An anipoint object.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Ensure every declared variable names a column that exists
}
\keyword{internal}
