% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anipoint_validate.R
\name{ensure_has_index}
\alias{ensure_has_index}
\title{Ensure the index column is present and numeric}
\usage{
ensure_has_index(data)
}
\arguments{
\item{data}{An anipoint object.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Ensure the index column is present and numeric
}
\keyword{internal}
