% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aniframe_is.R
\name{ensure_is_aniframe}
\alias{ensure_is_aniframe}
\title{Ensure object is an aniframe}
\usage{
ensure_is_aniframe(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
Error if not an aniframe
}
\description{
Guard form of \code{\link[=is_aniframe]{is_aniframe()}}, passed by any animovement frame class.
Functions that need coordinates should guard with
\code{\link[=ensure_is_anipoint]{ensure_is_anipoint()}} instead.
}
\examples{
af <- example_anipoint(n_obs = 3, n_individuals = 1, n_keypoints = 1)
# Passes silently, and errors otherwise
ensure_is_aniframe(af)

try(ensure_is_aniframe(data.frame(x = 1)))
}
