% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anipoint_is.R
\name{ensure_is_anipoint}
\alias{ensure_is_anipoint}
\title{Ensure object is an anipoint}
\usage{
ensure_is_anipoint(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
Error if not an anipoint
}
\description{
Ensure object is an anipoint
}
\examples{
af <- example_anipoint(n_obs = 3, n_individuals = 1, n_keypoints = 1)
# Passes silently, and errors otherwise
ensure_is_anipoint(af)

try(ensure_is_anipoint(data.frame(x = 1)))
}
