% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_system.R
\name{ensure_is_cartesian_1d}
\alias{ensure_is_cartesian_1d}
\title{Internal guard for 1-D Cartesian checks}
\usage{
ensure_is_cartesian_1d(data)
}
\arguments{
\item{data}{An aniframe.}
}
\description{
Internal guard for 1-D Cartesian checks
}
\examples{
af <- example_anipoint(n_obs = 3, n_individuals = 1, n_keypoints = 1)
try(ensure_is_cartesian_1d(af))
}
