% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_extents.R
\name{ensure_valid_axis_extents}
\alias{ensure_valid_axis_extents}
\title{Is this a usable map of axis roles to extents?}
\usage{
ensure_valid_axis_extents(extents)
}
\arguments{
\item{extents}{A proposed \code{axis_extents} value.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Is this a usable map of axis roles to extents?
}
\keyword{internal}
