% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_index.R
\name{ensure_valid_index_unit}
\alias{ensure_valid_index_unit}
\title{Ensure a unit for the index is a level of unit_time}
\usage{
ensure_valid_index_unit(unit, call = rlang::caller_env())
}
\arguments{
\item{unit}{The proposed unit, or \code{NULL}.}

\item{call}{The caller's environment, for the error.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Ensure a unit for the index is a level of unit_time
}
\keyword{internal}
