% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ensure_variables_character}
\alias{ensure_variables_character}
\title{Ensure a declaration is a character vector}
\usage{
ensure_variables_character(variables)
}
\arguments{
\item{variables}{Value supplied by the caller.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Ensure a declaration is a character vector
}
\keyword{internal}
