% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{frame_class}
\alias{frame_class}
\title{The frame class whose metadata schema applies to \code{data}}
\usage{
frame_class(data)
}
\description{
The frame class whose metadata schema applies to \code{data}
}
\keyword{internal}
