% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_extents.R
\name{get_axis_extents}
\alias{get_axis_extents}
\title{Get how far each axis runs}
\usage{
get_axis_extents(data)
}
\arguments{
\item{data}{An aniframe or anievent object.}
}
\value{
Named numeric vector, axis role to extent. Empty when the frame
declares none.
}
\description{
Get how far each axis runs
}
\examples{
af <- example_anipoint(n_obs = 3, n_individuals = 1, n_keypoints = 1)
get_axis_extents(af)

}
\seealso{
\code{\link[=set_axis_extents]{set_axis_extents()}}, \code{\link[=get_axis_directions]{get_axis_directions()}}
}
