% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{get_keys}
\alias{get_keys}
\title{The grouping columns of a frame}
\usage{
get_keys(data)
}
\arguments{
\item{data}{An aniframe.}
}
\value{
Character vector of column names.
}
\description{
The frame is grouped by its identity keys and temporal context keys:
\code{c(what$keys, when$keys)}. The index and an anievent's interval are never
grouping columns.
}
\examples{
af <- example_anipoint(n_obs = 3, n_individuals = 1, n_keypoints = 1)
get_keys(af)

}
\seealso{
\code{\link[=get_variables]{get_variables()}}
}
