% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_get.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata}
\usage{
get_metadata(data, fields = NULL)
}
\arguments{
\item{data}{An aniframe or anievent object.}

\item{fields}{If only specific metadata fields should be returned. A field
the object does not carry gives \code{NULL}; a name that is not a metadata
field at all is an error.}
}
\value{
The metadata associated with the object.
}
\description{
Get metadata
}
\examples{
af <- example_anipoint(n_obs = 3, n_individuals = 1, n_keypoints = 1)
names(get_metadata(af))

# A single field can be pulled out by name
get_metadata(af, 'sampling_rate')
}
