% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_accessors.R
\name{get_sampling_rate}
\alias{get_sampling_rate}
\title{The sampling rate, in Hz}
\usage{
get_sampling_rate(data)
}
\arguments{
\item{data}{An aniframe or anievent object.}
}
\value{
Numeric scalar, or \code{NA} when the rate is not recorded.
}
\description{
The sampling rate, in Hz
}
\examples{
af <- example_anipoint(n_obs = 3, n_individuals = 1, n_keypoints = 1)
get_sampling_rate(af)

}
\seealso{
\code{\link[=set_sampling_rate]{set_sampling_rate()}}
}
