% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_accessors.R
\name{get_unit_time}
\alias{get_unit_time}
\title{The unit the index or bout boundaries are in}
\usage{
get_unit_time(data)
}
\arguments{
\item{data}{An aniframe or anievent object.}
}
\value{
Length-one character vector.
}
\description{
The unit the index or bout boundaries are in
}
\examples{
af <- example_anipoint(n_obs = 3, n_individuals = 1, n_keypoints = 1)
get_unit_time(af)

}
\seealso{
\code{\link[=set_unit_time]{set_unit_time()}}
}
