% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anievent_is.R
\name{is_anievent}
\alias{is_anievent}
\title{Check if object is an anievent}
\usage{
is_anievent(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical: \code{TRUE} if \code{x} inherits from \code{anievent}.
}
\description{
Check if object is an anievent
}
\examples{
af <- example_anipoint(n_obs = 3, n_individuals = 1, n_keypoints = 1)
is_anievent(af)
}
