% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anijoint.R
\name{is_anijoint}
\alias{is_anijoint}
\alias{ensure_is_anijoint}
\title{Test whether an object is an anijoint}
\usage{
is_anijoint(x)

ensure_is_anijoint(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{is_anijoint()}: logical. \code{ensure_is_anijoint()}: errors if not.
}
\description{
Test whether an object is an anijoint
}
\examples{
af <- example_anipoint(n_obs = 3, n_individuals = 1) |>
  set_structure(example_structure())
is_anijoint(as_anijoint(af))
}
