% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anisegment.R
\name{is_anisegment}
\alias{is_anisegment}
\alias{ensure_is_anisegment}
\title{Test whether an object is an anisegment}
\usage{
is_anisegment(x)

ensure_is_anisegment(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{is_anisegment()}: logical. \code{ensure_is_anisegment()}: errors if
not.
}
\description{
Test whether an object is an anisegment
}
\examples{
af <- example_anipoint(n_obs = 3, n_individuals = 1) |>
  set_structure(example_structure())
is_anisegment(as_anisegment(af))
}
