% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anistructure.R
\name{is_anistructure}
\alias{is_anistructure}
\alias{ensure_is_anistructure}
\title{Test whether an object is an anistructure}
\usage{
is_anistructure(x)

ensure_is_anistructure(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{is_anistructure()}: logical. \code{ensure_is_anistructure()}: errors
if not.
}
\description{
Test whether an object is an anistructure
}
\examples{
is_anistructure(example_structure())
}
