% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_system.R
\name{is_cartesian_3d}
\alias{is_cartesian_3d}
\title{Test for a 3-D Cartesian coordinate system}
\usage{
is_cartesian_3d(data)
}
\arguments{
\item{data}{An anipoint.}
}
\value{
A logical value.
}
\description{
Test for a 3-D Cartesian coordinate system
}
\examples{
af <- example_anipoint(n_obs = 3, n_individuals = 1, n_keypoints = 1)
is_cartesian_3d(af)
}
