% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tree.R
\name{is_nested_metadata}
\alias{is_nested_metadata}
\title{Is this metadata list in the nested (category) layout?}
\usage{
is_nested_metadata(md)
}
\arguments{
\item{md}{A metadata list.}
}
\value{
Logical scalar.
}
\description{
Is this metadata list in the nested (category) layout?
}
\keyword{internal}
