% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_set.R
\name{is_no_fixed_rate}
\alias{is_no_fixed_rate}
\title{Is this the marker for a device with no fixed rate?}
\usage{
is_no_fixed_rate(x)
}
\arguments{
\item{x}{A field's value.}
}
\value{
Logical scalar.
}
\description{
Is this the marker for a device with no fixed rate?
}
\keyword{internal}
