% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_time.R
\name{key_group_ids}
\alias{key_group_ids}
\title{An integer naming each row's identity and temporal context}
\usage{
key_group_ids(data)
}
\arguments{
\item{data}{An anipoint or anievent.}
}
\value{
Integer vector, one per row.
}
\description{
An integer naming each row's identity and temporal context
}
\keyword{internal}
