% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_directions.R
\name{list_axis_directions}
\alias{list_axis_directions}
\title{The directions an axis can point, as seen from the recording viewpoint}
\usage{
list_axis_directions()
}
\value{
Character vector of the permitted directions.
}
\description{
The directions an axis can point, as seen from the recording viewpoint
}
\keyword{internal}
