% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axes.R
\name{list_axis_role_sets}
\alias{list_axis_role_sets}
\title{The axis role sets that define each coordinate system}
\usage{
list_axis_role_sets()
}
\value{
Named list mapping a comma-separated sorted role set to the
coordinate system it defines.
}
\description{
The axis role sets that define each coordinate system
}
\keyword{internal}
