% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_directions.R
\name{list_direction_pairs}
\alias{list_direction_pairs}
\title{Which opposed pair each direction belongs to}
\usage{
list_direction_pairs()
}
\value{
Named character vector, direction to pair.
}
\description{
Which opposed pair each direction belongs to
}
\keyword{internal}
