% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_directions.R
\name{list_linear_axis_roles}
\alias{list_linear_axis_roles}
\title{The axis roles that can point somewhere (the Cartesian ones)}
\usage{
list_linear_axis_roles()
}
\value{
Character vector of axis roles.
}
\description{
The axis roles that can point somewhere (the Cartesian ones)
}
\keyword{internal}
