% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tree.R
\name{list_metadata_categories}
\alias{list_metadata_categories}
\title{The metadata categories}
\usage{
list_metadata_categories()
}
\value{
Character vector of category names.
}
\description{
The metadata categories
}
\keyword{internal}
