% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_print.R
\name{print_metadata_leaves}
\alias{print_metadata_leaves}
\title{Render one category's leaves as aligned name/type/value lines}
\usage{
print_metadata_leaves(x)
}
\arguments{
\item{x}{A named list of leaf values.}
}
\description{
Render one category's leaves as aligned name/type/value lines
}
\keyword{internal}
