% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_print.R
\name{print_metadata_structure}
\alias{print_metadata_structure}
\title{Render the structure category: one line per structure}
\usage{
print_metadata_structure(structure)
}
\arguments{
\item{structure}{The structure list.}
}
\description{
Render the structure category: one line per structure
}
\keyword{internal}
