% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angles.R
\name{rad_to_deg}
\alias{rad_to_deg}
\title{Convert radians to degrees}
\usage{
rad_to_deg(x)
}
\arguments{
\item{x}{Numeric vector of angles (radians).}
}
\value{
Numeric vector of angles expressed in degrees.
}
\description{
Convert radians to degrees
}
\examples{
rad_to_deg(pi)
rad_to_deg(c(0, pi / 2, pi))

}
\seealso{
Other angle utilities:
\code{\link[=angle_to_rad]{angle_to_rad()}},
\code{\link[=deg_to_rad]{deg_to_rad()}},
\code{\link[=unwrap_angle]{unwrap_angle()}},
\code{\link[=wrap_angle]{wrap_angle()}}
}
\concept{angle utilities}
