% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anisegment.R
\name{restructure_anisegment}
\alias{restructure_anisegment}
\title{Order, group and declare an anisegment}
\usage{
restructure_anisegment(data, md)
}
\arguments{
\item{data}{An anisegment.}

\item{md}{Its complete metadata.}
}
\value{
\code{data}, restructured, with \code{md} attached.
}
\description{
Order, group and declare an anisegment
}
\keyword{internal}
