% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anisegment.R
\name{restructure_value_frame}
\alias{restructure_value_frame}
\title{Order, group and declare a segment or joint frame}
\usage{
restructure_value_frame(data, md)
}
\arguments{
\item{data}{An anisegment or anijoint.}

\item{md}{Its complete metadata.}
}
\value{
\code{data}, restructured, with \code{md} attached.
}
\description{
Order, group and declare a segment or joint frame
}
\keyword{internal}
