% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anievent_as.R
\name{standardise_anievent_cols}
\alias{standardise_anievent_cols}
\title{Standardise column types for an anievent}
\usage{
standardise_anievent_cols(data, variables_what, variables_when)
}
\arguments{
\item{data}{Data frame to standardise.}

\item{variables_what}{Identity variable names.}

\item{variables_when}{Temporal variable names, including \code{start}/\code{stop}.}
}
\value{
Data frame with standardised column types.
}
\description{
Standardise column types for an anievent
}
\keyword{internal}
