% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure.R
\name{standardise_anipoint_cols}
\alias{standardise_anipoint_cols}
\title{Standardize column types for anipoint}
\usage{
standardise_anipoint_cols(
  data,
  variables_what,
  variables_when,
  variables_where,
  index = "time"
)
}
\arguments{
\item{data}{Data frame to standardise.}

\item{variables_what}{Identity variable names.}

\item{variables_when}{Temporal variable names.}

\item{variables_where}{Spatial variable names.}

\item{index}{The index column, which stays numeric. The temporal
context variables are made categorical.}
}
\value{
Data frame with standardised column types.
}
\description{
Converts character identity and temporal variables to factors.
Converts numeric identity and temporal variables (except the index) to
integers.
Spatial variables are converted to numeric.
}
\keyword{internal}
