% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ensure_has_declared_cols}
\alias{ensure_has_declared_cols}
\title{Ensure declared columns are present}
\usage{
ensure_has_declared_cols(data, cols, role)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{Character vector of declared column names.}

\item{role}{One of \code{"what"}, \code{"when"}, \code{"where"}.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Ensure declared columns are present
}
\keyword{internal}
