% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_system.R
\name{is_spherical}
\alias{is_spherical}
\title{Test whether an anipoint uses a spherical coordinate system}
\usage{
is_spherical(data)
}
\arguments{
\item{data}{An anipoint.}
}
\value{
A logical value.
}
\description{
Test whether an anipoint uses a spherical coordinate system
}
\examples{
af <- example_anipoint(n_obs = 3, n_individuals = 1, n_keypoints = 1)
is_spherical(af)
}
