% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anipoint_new.R
\name{new_anipoint}
\alias{new_anipoint}
\title{Create a new anipoint object (internal constructor)}
\usage{
new_anipoint(x)
}
\arguments{
\item{x}{A data frame to convert to anipoint}
}
\value{
An anipoint object
}
\description{
Create a new anipoint object (internal constructor)
}
\keyword{internal}
