% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_time.R
\name{resolve_unit_time_calibration}
\alias{resolve_unit_time_calibration}
\title{The multiplier for a unit_time conversion}
\usage{
resolve_unit_time_calibration(data, to_unit, calibration_factor)
}
\description{
The multiplier for a unit_time conversion
}
\keyword{internal}
