% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anistructure.R
\name{validate_anistructure}
\alias{validate_anistructure}
\title{Check that a structure is internally consistent}
\usage{
validate_anistructure(x)
}
\arguments{
\item{x}{An \code{anistructure}.}
}
\value{
\code{x}, invisibly; errors naming the first problem otherwise.
}
\description{
Check that a structure is internally consistent
}
\examples{
validate_anistructure(example_structure())
}
