% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anipoint_validate.R
\name{warn_coordinate_system_drift}
\alias{warn_coordinate_system_drift}
\title{Warn when coordinate_system no longer matches variables_where}
\usage{
warn_coordinate_system_drift(data)
}
\arguments{
\item{data}{An anipoint object.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Assumes \code{\link[=ensure_is_spatial]{ensure_is_spatial()}} has already passed.
}
\keyword{internal}
